% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{combine.matrices.by.rowname.intersect}
\alias{combine.matrices.by.rowname.intersect}
\title{combine.matrices.by.rowname.intersect}
\usage{
combine.matrices.by.rowname.intersect(matrix1, matrix2, k = 2)
}
\arguments{
\item{matrix1}{PARAM_DESCRIPTION}

\item{matrix2}{PARAM_DESCRIPTION}

\item{k}{PARAM_DESCRIPTION, Default: 2}
}
\description{
Combine matrices by rownames intersect.
}
