% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{FALSE.unless}
\alias{FALSE.unless}
\title{FALSE.unless}
\usage{
FALSE.unless(NameOfaVariable = "VarName")
}
\arguments{
\item{NameOfaVariable}{Name of a possibly defined variable to be tested.}
}
\description{
Return FALSE unless the variable is defined. If defined, it returns the value of the variable.
}
\examples{
FALSE.unless("xsadasf32")
Num <- 22
FALSE.unless("Num")
FALSE.unless("c")
}
