% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{filter_survival_length}
\alias{filter_survival_length}
\title{filter_survival_length}
\usage{
filter_survival_length(length_new, length_old, prepend = "")
}
\arguments{
\item{length_new}{The number of elements that survived the filter.}

\item{length_old}{The total number of elements.}

\item{prepend}{A string to prepend to the sentence.}
}
\value{
A string.
}
\description{
Parse a sentence reporting the \% of filter survival.
}
