% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{jjpegA4}
\alias{jjpegA4}
\title{jjpegA4}
\usage{
jjpegA4(filename, r = 225, q = 90)
}
\arguments{
\item{filename}{PARAM_DESCRIPTION}

\item{r}{PARAM_DESCRIPTION, Default: 225}

\item{q}{PARAM_DESCRIPTION, Default: 90}
}
\description{
Setup an A4 size jpeg
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
