% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{llogit}
\alias{llogit}
\title{llogit}
\usage{
llogit(...)
}
\arguments{
\item{...}{Variables (strings, vectors) to be collapsed consecutively.}
}
\description{
Collapse a sentence from any variables passed to the function with white spaces.
llogit() writes it to your markdown report file, if the "path_of_report" variable is defined.
It does not print the sentence to the screen.
}
\examples{
MyFriends <- c("Peter", "Bence")
llogit("My friends are: ", MyFriends)
}
