% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{llprint}
\alias{llprint}
\title{llprint}
\usage{
llprint(...)
}
\arguments{
\item{...}{Variables (strings, vectors) to be collapsed consecutively.}
}
\description{
Collapse a sentence from any variables passed to the function with white spaces.
Print the sentence to the screen and write it to your markdown report file,
if the "path_of_report" variable is defined.
}
\examples{
MyFriends <- c("Peter", "Bence")
llprint("My friends are: ", MyFriends)
}
