% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{lookup}
\alias{lookup}
\title{lookup}
\usage{
lookup(needle, haystack, exact = TRUE, report = FALSE)
}
\arguments{
\item{needle}{A vector of values to look for.}

\item{haystack}{A vector of values to search in.}

\item{exact}{Logical. Whether to do an exact match or a partial match.}

\item{report}{Logical. Whether to print a report of the results.}
}
\value{
A list with the results of the lookup.
}
\description{
Awesome pattern matching for a set of values in another set of values. Returns a list with all kinds of results.
}
