% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{md.import}
\alias{md.import}
\title{md.import}
\usage{
md.import(from.file, to.file = ww.set.path_of_report())
}
\arguments{
\item{from.file}{File to be appended at the current last line of the report}

\item{to.file}{The report file. Defined as "path_of_report" by default,
which is set by the "setup_MarkdownReports" function.}
}
\description{
Import and concatenate an external markdown or text file into the report
}
\examples{
path_of_report <- ww.set.path_of_report()
llprint("Hello")
# md.import(path_of_report)

}
