% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{md.LogSettingsFromList}
\alias{md.LogSettingsFromList}
\title{md.LogSettingsFromList}
\usage{
md.LogSettingsFromList(parameterlist, maxlen = 20)
}
\arguments{
\item{parameterlist}{List of Parameters}

\item{maxlen}{Maximum length of entries in a parameter list element. Default: 20}
}
\description{
Log the parameters & settings used in the script and stored in a list, in a table format
in the report.
}
\examples{
md.LogSettingsFromList(parameterlist = list("min" = 4, "method" = "pearson", "max" = 10))
}
