% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{md.write.as.list}
\alias{md.write.as.list}
\title{md.write.as.list}
\usage{
md.write.as.list(
  vector = 1:3,
  h = 4,
  numbered = FALSE,
  path_of_report = ww.set.path_of_report(),
  ...
)
}
\arguments{
\item{vector}{Vector to be written as a list}

\item{h}{Level of header above the list.}

\item{numbered}{TRUE = Numbered list, FALSE = unordered list is written}

\item{path_of_report}{Path to the report file. Default: \code{ww.set.path_of_report()}.}

\item{...}{Additional parameters}
}
\description{
Writes a vector as a (numbered) list into the report file.
}
\examples{
md.write.as.list()
}
