% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{TRUE.unless}
\alias{TRUE.unless}
\title{TRUE.unless}
\usage{
TRUE.unless(NameOfaVariable = "VarName")
}
\arguments{
\item{NameOfaVariable}{Name of a possibly defined variable to be tested.}
}
\description{
Return TRUE unless the variable is defined. If defined, it returns the value of the variable.
}
\examples{
TRUE.unless("xsadasf32"); Num = 22; TRUE.unless("Num"); TRUE.unless("c")
}
