% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{try_err2warn}
\alias{try_err2warn}
\title{try_err2warn}
\usage{
try_err2warn(expr, fallback, warn = NULL)
}
\arguments{
\item{expr}{Expression to be evaluated.}

\item{fallback}{Value to return if \code{expr} errors.}

\item{warn}{Optional character string prepended to the warning message.}
}
\description{
Evaluate an expression and return its value.
If an error occurs, convert the error into an immediate warning and return a fallback value.
Execution is never stopped.
}
\examples{
try_err2warn(stop("fail"), 1)

x <- 2
try_err2warn(x + 1, x)

try_err2warn(stop("boom"), 0, "Computation failed")
}
