% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{wcolorize}
\alias{wcolorize}
\title{wcolorize}
\usage{
wcolorize(
  vector = c(1, 1, 1:6),
  RColorBrewerSet = FALSE,
  ReturnCategoriesToo = FALSE,
  show = FALSE,
  randomize = FALSE,
  set = c(FALSE, "rich", "heat.colors", "terrain.colors", "topo.colors", "matlab",
    "rainbow")[1]
)
}
\arguments{
\item{vector}{A vector with categories, can be numbers or strings}

\item{RColorBrewerSet}{Use one of the RColorBrewer color sets? Provide that name}

\item{ReturnCategoriesToo}{Return unique Categories. See example.}

\item{show}{Show generated color palette}

\item{randomize}{Randomize colors}

\item{set}{Color palette for base
("heat.colors", "terrain.colors", "topo.colors", "rainbow"),
or "rich" for gplots::rich.colors, or  "matlab" for colorRamps::matlab.like.}

\item{prefix}{String before the subscript.}

\item{subscr}{Subscripted text.}

\item{quantity}{String in brackets after the subscript, eg.: log2(read count).}
}
\description{
Generate color palettes. Input: a vector with categories, can be numbers or strings.
Handles repeating values. Output: color vector of equal length as input.
Optionally it can ouput a list where an extra element lists the
categories (simply using unique would remove the names). See example.
Some color scale depend on packages "colorRamps", or "gplots".
}
\details{
subscript_in_plots

Returns a formatted string that you feed to main, xlab or ylab parameters of a plot
Create an expression with subscript for axis labels.
Parsed when provided to xlab or ylab of a function.
}
\examples{
plot (1, 1, xlab = subscript_in_plots(subscr = 10, quantity = "read count"),
 ylab = subscript_in_plots())
wcolorize (vector = c(1, 1, 1:6), ReturnCategoriesToo = TRUE, show = TRUE)
}
