% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{ww.set.file.extension}
\alias{ww.set.file.extension}
\title{ww.set.file.extension}
\usage{
ww.set.file.extension(
  also_pdf = also.pdf,
  def = "png",
  global_setting = "b.def.ext"
)
}
\arguments{
\item{also_pdf}{Save plot in both png and pdf formats.}

\item{def}{Default file extension: png}

\item{global_setting}{global file extension setting stored in a global variable. Detault: 'b.def.ext'}
}
\description{
Internal function. Sets file extension for ggExpress plotting functions.
}
