% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{ww.set.mdlink}
\alias{ww.set.mdlink}
\title{ww.set.mdlink}
\usage{
ww.set.mdlink(NameOfaVariable = "b.mdlink", def = FALSE)
}
\arguments{
\item{NameOfaVariable}{Name of a possibly defined variable to be tested.}

\item{def}{Default return value}
}
\description{
Internal function. Inserts a markdown link to the image
(created by the wplot* function that calls this function) only if 'path_of_report' is defined
and 'b.mdlink' is TRUE.
}
\examples{
ww.set.mdlink() # It is an internal function, not intended for manual use.
}
