% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{ww.variable.and.path.exists}
\alias{ww.variable.and.path.exists}
\title{ww.variable.and.path.exists}
\usage{
ww.variable.and.path.exists(path = path_of_report, alt.message = NULL)

ww.variable.and.path.exists(path = path_of_report, alt.message = NULL)
}
\arguments{
\item{path}{A variable name that might not exist and might point to a non-existent directory.}

\item{alt.message}{Alternative message if the variable + path does not exist. FALSE or string.}
}
\description{
Check if a variable name is defined and, if so, does the path (to a file) stored in that
variable point to an existing directory?

Check if a variable name is defined and, if so, does the path (to a file) stored in that
variable point to an existing directory?
}
\examples{
ww.variable.and.path.exists(path = B, alt.message = "Hello, your path/var does not exist.")
ww.variable.and.path.exists(path = B, alt.message = "Hello, your path/var does not exist.")
}
