% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{FALSE.unless}
\alias{FALSE.unless}
\title{FALSE.unless}
\usage{
FALSE.unless(NameOfaVariable, v = TRUE)
}
\arguments{
\item{NameOfaVariable}{Variable to test, given as a bare name or a character string. Must not be \code{NULL}.}

\item{v}{verbose}
}
\description{
Return FALSE unless the variable is defined. If defined, it returns the value of the variable.
}
\examples{
FALSE.unless(xsadasf32)
FALSE.unless("xsadasf32")
Num <- 22
FALSE.unless(Num)
FALSE.unless("c")
}
