% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{jjpegA4}
\alias{jjpegA4}
\title{jjpegA4}
\usage{
jjpegA4(filename, r = 225, q = 90, w = 8.27, h = 11.69)
}
\arguments{
\item{filename}{The filename of the jpeg file to create.}

\item{r}{The resolution of the jpeg file.}

\item{q}{The quality of the jpeg file.}

\item{w}{Width}

\item{h}{Height}
}
\description{
Setup an A4 size jpeg.
}
\examples{
\dontrun{
if(interactive()){
 jjpegA4(filename = "my_plot.jpeg", r = 225, q = 90)
 }
}

}
