% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{ww.set.file.extension}
\alias{ww.set.file.extension}
\title{ww.set.file.extension}
\usage{
ww.set.file.extension(
  also.pdf = also.pdf,
  def = "png",
  global.setting = "b.def.ext"
)
}
\arguments{
\item{also.pdf}{Save plot in both png and pdf formats?}

\item{def}{Default file extension: png}

\item{global.setting}{global file extension setting stored in a global variable. Detault: 'b.def.ext'}
}
\description{
Internal function. Creates automatic plot and file-names.
}
