% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownHelpers.R
\name{ww.set.file.extension}
\alias{ww.set.file.extension}
\title{ww.set.file.extension}
\usage{
ww.set.file.extension(global_setting = "b.def.ext", default = "png", also_pdf)
}
\arguments{

\item{global_setting}{Global file extension setting stored in a global variable. Default: 'b.def.ext'}

\item{default}{Default file extension: png}

\item{also_pdf}{Save plot in both PNG and PDF formats.}
}
\description{
Internal function. Sets file extension for ggExpress plotting functions.
}
